Replace create page content
